Add transitions to all pages
