+++ src/components/AppContent/EducationPage/EducationCard/EducationCard.styles.ts
