+++ src/components/AppContent/EducationPage/EducationCard/index.ts
@@ -1,0 +1,3 @@
+import EducationCard from './EducationCard';
+
+export default EducationCard;
