+++ src/components/AppContent/ExperiencePage/ExperienceCard/ExperienceCard.styles.ts
