+++ src/components/AppContent/ExperiencePage/ExperienceCard/index.ts
@@ -1,0 +1,3 @@
+import ExperienceCard from './ExperienceCard';
+
+export default ExperienceCard;
