+++ src/components/AppContent/HBVResearchPage/HBVResearchCard/HBVResearchCard.styles.ts
