+++ src/components/AppContent/HBVResearchPage/HBVResearchCard/index.ts
@@ -1,0 +1,3 @@
+import HBVResearchCard from './HBVResearchCard';
+
+export default HBVResearchCard;
