+++ src/components/AppContent/HolisticOfficePage/ArchitecturePaper/ArchitecturePaper.tsx
@@ -1,4 +1,4 @@
-import { Paper, Typography } from '@material-ui/core';
+import { Grow, Paper, Typography } from '@material-ui/core';
 import { map } from 'lodash';
 import React from 'react';
 import { ArchitectureCategory } from 'types/holisticOffice';
@@ -14,20 +14,22 @@ const ArchitecturePaper: React.FC<ArchitecturePaperProps> = ({
 }: ArchitecturePaperProps) => {
     const classes = architecturePaperStyles();
     return (
-        <Paper className={classes.architecturePaper} elevation={5}>
-            <Typography variant="h5" align="center">
-                {title}
-            </Typography>
-            <Typography>
-                <ul>
-                    {map(mappings, (value: string, key: string) => (
-                        <li key={key}>
-                            <b>{key}:</b> {value}
-                        </li>
-                    ))}
-                </ul>
-            </Typography>
-        </Paper>
+        <Grow in>
+            <Paper className={classes.architecturePaper} elevation={5}>
+                <Typography variant="h5" align="center">
+                    {title}
+                </Typography>
+                <Typography>
+                    <ul>
+                        {map(mappings, (value: string, key: string) => (
+                            <li key={key}>
+                                <b>{key}:</b> {value}
+                            </li>
+                        ))}
+                    </ul>
+                </Typography>
+            </Paper>
+        </Grow>
     );
 };
