+++ src/components/AppContent/MartialArtsPage/MartialArtsPage.tsx
@@ -1,4 +1,4 @@
-import { Container, Grid, Typography } from '@material-ui/core';
+import { Container, Fade, Grid, Grow, Typography } from '@material-ui/core';
 import { useTheme } from '@material-ui/core/styles';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
 import clsx from 'clsx';
@@ -20,41 +20,51 @@ const MartialArtsPage: React.FC<MartialArtsPageProps> = ({ martialArtsStyle }: M
     const martialArt: MartialArt | undefined = useMartialArt(martialArtsStyle);
     const theme = useTheme();
     const isLarge = useMediaQuery(theme.breakpoints.up('lg'));
-    return martialArt ? (
+    return (
         <div className={clsx(shared.bg, classes.bg)}>
-            <Container maxWidth="lg" className={classes.contentBg}>
-                <div className={classes.pageHeading}>
-                    {isLarge && <img src={martialArt.logoUrl} alt={`${martialArt.style}_logo`} />}
-                    <Typography variant="h1" align="center" className={classes.title}>
-                        {martialArt.styleName}
-                    </Typography>
-                    {isLarge && <img src={martialArt.logoUrl} alt={`${martialArt.style}_logo`} />}
-                </div>
-                <Typography variant="h4" align="center" className={classes.rank}>
-                    Current Rank: {martialArt.blackBeltRank} 단
-                </Typography>
-                <Grid container spacing={3} className={classes.pageContent}>
-                    <Grid xs={12} lg={6}>
-                        <Typography paragraph className={classes.introduction}>
-                            {martialArt.introduction}
+            {martialArt ? (
+                <Fade in>
+                    <Container maxWidth="lg" className={classes.contentBg}>
+                        <div className={classes.pageHeading}>
+                            {isLarge && <img src={martialArt.logoUrl} alt={`${martialArt.style}_logo`} />}
+                            <Typography variant="h1" align="center" className={classes.title}>
+                                {martialArt.styleName}
+                            </Typography>
+                            {isLarge && <img src={martialArt.logoUrl} alt={`${martialArt.style}_logo`} />}
+                        </div>
+                        <Typography variant="h4" align="center" className={classes.rank}>
+                            Current Rank: {martialArt.blackBeltRank} 단
                         </Typography>
-                        <Typography paragraph>{martialArt.biography}</Typography>
-                    </Grid>
-                    <Grid xs={12} lg={6} className={classes.picContainer}>
-                        <figure className={classes.pic}>
-                            <img className={classes.picImg} src={martialArt.imgUrl} alt={`${martialArt.style}_img`} />
-                            <figcaption className={classes.picCaption}>
-                                <Typography variant="subtitle1" align="center">
-                                    {martialArt.imgCaption}
+                        <Grid container spacing={3} className={classes.pageContent}>
+                            <Grid xs={12} lg={6}>
+                                <Typography paragraph className={classes.introduction}>
+                                    {martialArt.introduction}
                                 </Typography>
-                            </figcaption>
-                        </figure>
-                    </Grid>
-                </Grid>
-                <StudioGrid studios={martialArt.studios} />
-            </Container>
+                                <Typography paragraph>{martialArt.biography}</Typography>
+                            </Grid>
+                            <Grid xs={12} lg={6} className={classes.picContainer}>
+                                <Grow in>
+                                    <figure className={classes.pic}>
+                                        <img
+                                            className={classes.picImg}
+                                            src={martialArt.imgUrl}
+                                            alt={`${martialArt.style}_img`}
+                                        />
+                                        <figcaption className={classes.picCaption}>
+                                            <Typography variant="subtitle1" align="center">
+                                                {martialArt.imgCaption}
+                                            </Typography>
+                                        </figcaption>
+                                    </figure>
+                                </Grow>
+                            </Grid>
+                        </Grid>
+                        <StudioGrid studios={martialArt.studios} />
+                    </Container>
+                </Fade>
+            ) : null}
         </div>
-    ) : null;
+    );
 };
 
 export default MartialArtsPage;
