+++ src/components/AppContent/ProjectsPage/ProjectCard/index.ts
@@ -1,0 +1,3 @@
+import ProjectCard from './ProjectCard';
+
+export default ProjectCard;
